requirements.txt
# Recommended to install in virtual environment ./.venv
# python3 -m pip install -r requirements.txt
openssl-python
pycryptodome
rsa
tinyec
python-secrets